Add image upload functionality with multer
A new feature to upload image files is added by integrating multer package.
Multer is a Node.js middleware added to handle multipart/form-data which is primarily used for file uploads. The new feature includes routing and middleware to handle typical single file upload use case. Multer package is added to package.json and package-lock.json. The image files are stored in a new directory named 'uploads'. This feature required modifications in index.js and the addition of several dependencies and sub-dependencies.